fix all lint issues
