+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-native-masked-view/masked-view": "^0.2.8",
+    "@react-navigation/elements": "^1.3.17",
     "@react-navigation/native": "^6.1.4",
     "@react-navigation/native-stack": "^6.9.10",
     "axios": "^1.3.3",
