+++ src/components/notes-card/NoteCard.js
@@ -1,4 +1,4 @@
-import { StyleSheet, View, Pressable } from 'react-native';
+import { View, Pressable } from 'react-native';
 import { Headline, Subhead, Caption1 } from 'react-native-ios-kit';
 import styles from './styles';
 
@@ -11,8 +11,8 @@ const NotesCard = ({ note, style, onPress }) => {
                 style,
             ]}>
             <View style={{ flex: 1, justifyContent: 'space-between' }}>
-                <Headline>{note?.title}</Headline>
-                <Subhead>{note?.body}</Subhead>
+                <Headline numberOfLines={1}>{note?.title}</Headline>
+                <Subhead numberOfLines={2}>{note?.body}</Subhead>
                 <Caption1>{note?.date}</Caption1>
             </View>
         </Pressable>
