+++ src/components/notes-card/SkeletonCard.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import SkeletonPlaceholder from 'react-native-skeleton-placeholder';
 import { useWindowDimensions, View } from 'react-native';
 import styles from './styles';
