+++ src/screens/home/SkeletonPage.js
@@ -1,13 +1,14 @@
 
 import React from 'react';
-import SkeletonComponent from '../../components/notes-card/SkeletonCard';
 import { View } from 'react-native';
+import SkeletonComponent from '../../components/notes-card/SkeletonCard';
+import styles from './styles';
 
 class SkeletonPage extends React.PureComponent {
     render() {
         return (<>
             {
-                [...Array(5)].map((_, index) => (<View key={index} style={{ paddingBottom: 20 }}>
+                ['c1', 'c2', 'c3'].map((val) => (<View key={val} style={styles.skeleton}>
                     <SkeletonComponent />
                 </View>))
             }
