+++ src/screens/home/index.js
@@ -1,2 +1,3 @@
 import HomeScreen from "./HomeScreen";
+
 export default HomeScreen;
