+++ src/screens/note/NoteTextField.js
@@ -1,15 +1,12 @@
 import { TextField } from 'react-native-ios-kit';
-import { Colors } from 'react-native-ios-kit';
 
+/* eslint-disable react/jsx-props-no-spreading */
 const NoteTextField = ({ placeholder, value, onValueChange, inputStyle, multiline, ...rest }) => {
     return (<TextField
         placeholder={placeholder}
         value={value}
         onValueChange={onValueChange}
         clearButton={false}
-        theme={{
-            dividerColor: Colors.white
-        }}
         inputStyle={inputStyle}
         multiline={multiline}
         {...rest}
