+++ src/screens/note/index.js
@@ -1,2 +1,3 @@
 import NoteScreen from "./NoteScreen";
+
 export default NoteScreen;
